+++ Editor/InspectorInjectionBackendFactory.cs
@@ -1,0 +1,12 @@
+#if UNITY_EDITOR
+namespace Jeomseon.Attribute.Editor
+{
+    internal static class InspectorInjectionBackendFactory
+    {
+        public static IInspectorInjectionBackend Create()
+        {
+            return new Unity6InspectorInjectionBackend();
+        }
+    }
+}
+#endif