+++ Runtime/PackageArchitectureNotes.cs
@@ -1,9 +1,0 @@
-namespace Jeomseon.Attribute
-{
-    internal static class PackageArchitectureNotes
-    {
-        // TODO(api): Attribute 선언은 외부 선택 패키지에 의존하지 않도록 유지하고,
-        // 기능별 Attribute는 해당 기능 패키지에서 소유하도록 분리합니다.
-        // TODO(editor): PropertyDrawer 구현은 EditorToolkit에서 제공하되 Unity 공식 Editor API만 사용합니다.
-    }
-}